+++ packages/control/src/migrations/0001_control_tables.ts
@@ -87,11 +87,9 @@ const statements = [
  * @category migrations
  * @since 0.1.0
  */
-const initial: Effect.Effect<void, unknown, SqlClient.SqlClient> = Effect.gen(function*() {
+export const initial: Effect.Effect<void, unknown, SqlClient.SqlClient> = Effect.gen(function*() {
   const sql = yield* SqlClient.SqlClient
   for (const statement of statements) {
     yield* sql.unsafe(statement)
   }
 })
-
-export default initial